% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDetP.R
\name{analyzeDetP}
\alias{analyzeDetP}
\title{Analyze detection p-values}
\usage{
analyzeDetP(detP, pDat, cpg_threshold = 0.01)
}
\arguments{
\item{detP}{A p by n dataframe matrix of detection p values. For example from
minfi::detectionP(rgset).}

\item{pDat}{A n by m dataframe/tibble of metadata.}

\item{cpg_threshold}{A numeric value x, where 0 < x < 1, that specifies the 
detection p value threshold for a poor quality probe.}
}
\value{
List of length two:
\itemize{
  \item fp: p by n dataframe with elements where 1's are placed when an
  observation failed cpg_threshold, and 0's otherwise.
  \item pDat: n by m + 1 dataframe containing a column ndetp corresponding
  to the number of probes that failed in sample i, where i = 1, 2, ..., n.
}
}
\description{
\code{analyzeDetP} Applies given thresholds to detection p value matrix to 
identify poor quality probes and samples.
}
\examples{
library(minfiData)
library(dplyr)
data(RGsetEx)

x <- analyzeDetP(detP = detectionP(RGsetEx), 
                 pDat = as_tibble(pData(RGsetEx), 
                 cpg_threshold = 0.01))
}
