% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH_mprice}
\alias{CH_mprice}
\title{List of XY- and autosomal cross-hybridizing probes from mprice et al. (2013)}
\format{A character vector of CpG names with length 2083}
\source{
Price, E. M., Cotton, A. M., Lam, L. L., Farré, P., Emberly, E.,
Brown, C. J., … Kobor, M. S. (2013). Additional annotation enhances potential
for biologically-relevant analysis of the Illumina Infinium
HumanMethylation450 BeadChip array. Epigenetics & Chromatin, 6, 4.
http://doi.org/10.1186/1756-8935-6-4
}
\usage{
CH_mprice
}
\description{
A character vector of 2083 CpG names that were identified to be able to cross
hybridize to other locations in the genome (both XY and autosomal).
}
\keyword{datasets}
