% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chen_CH}
\alias{chen_CH}
\title{List of cross-hybridizing probes from Chen et al. (2013)}
\format{A character vector of CpG names with length 29233}
\source{
Price, E. M., Cotton, A. M., Lam, L. L., Farré, P., Emberly, E.,
Brown, C. J., … Kobor, M. S. (2013). Additional annotation enhances potential
for biologically-relevant analysis of the Illumina Infinium
HumanMethylation450 BeadChip array. Epigenetics & Chromatin, 6, 4.
http://doi.org/10.1186/1756-8935-6-4

http://www.sickkids.ca/MS-Office-Files/Research/Weksberg%20Lab/'
file: '48639-non-specific-probes-Illumina450k.xlsx'
}
\usage{
chen_CH
}
\description{
A character vector of 29233 CpG names that were identified to be able to 
cross hybridize to other locations in the genome.
}
\keyword{datasets}
