% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulative}
\alias{cumulative}
\title{The cumulative amount of variance explained in each PC while adjusting for
other meta variables.}
\usage{
cumulative(pc_matrix, pDat, response)
}
\arguments{
\item{pc_matrix}{A rotated p x n dataframe, where p are the PC scores
(rotated values) and n are samples.}

\item{pDat}{A n x m dataframe, containing both the metavariables to be
adjusted and the response variable}

\item{response}{The metavariable to be used as the dependent variable.}
}
\value{
returns a list of Rsquared values from each linear model
}
\description{
\code{cumulative} progressively adds more and more PCs into a linear model,
while adjusting for other meta variables in \code{pDat}.
}
\details{
Creates progressively longer glm models of the form, while
adjusting for other columns in pDat, such as sex:
\itemize{
  \item response ~ pc1 + [Sex]
  \item response ~ pc1 + pc2 + [Sex]
  \item response ~ pc1 + pc2 + ...  + [Sex]
}
}
