% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatPCA.R
\name{heatPCA}
\alias{heatPCA}
\title{Create a heatmap of association between PCs and clinical variables}
\usage{
heatPCA(pc_matrix, pDat, metric = "Rsquared", labels = NULL,
  returnMat = F)
}
\arguments{
\item{pc_matrix}{A rotated p x n dataframe, where p are the PC scores
(rotated values) and n are samples.}

\item{pDat}{A n x m dataframe, where m is the number of clinical / meta
variables to be tested for association with the PCs.}

\item{metric}{Defaults to 'Rsquared'. 'P-value' not implemented yet.}

\item{returnMat}{A logical indicating if a dataframe of association values
should be returned}
}
\value{
A dataframe of association values between PCs and each meta variable
}
\description{
\code{heatPCA} tests association between principal components calculated on
gene expression or DNAm data and clinical variables, the plots these on a
heatmap.
}
\details{
Each PC supplied (I recommend only doing the top 10-20 for
readability) is tested for their correlation / association with each PC by
computing a linear model between each PC (dependent variable) and each
meta variable (dependent variable). The results are then plotted using
pheatmap::pheatmap and RColorBrewer.
##TODO:: implement % variance explained.
##TODO:: implement return p-values
}
