% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invar_Bl}
\alias{invar_Bl}
\title{Probes identified as invariable in indepedent blood datasets}
\format{A character vector of probe names with length 114204}
\source{
Edgar R.D., Jones M.J., Robinson W.P., Kobor M.S. (2017) An 
empirically driven data reduction method on the human 450K methylation array 
to remove tissue specific non-variable CpGs. Clin. Epigenet ., 9, 11.
https://github.com/redgar598/Tissue_Nonvariable_450K_CpGs
}
\usage{
invar_Bl
}
\description{
A character vector of 114204 probe names identified as nonvariable between
multiple BLOOD datasets.
}
\keyword{datasets}
