% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invar_Bu}
\alias{invar_Bu}
\title{Probes identified as invariable in indepedent buccal datasets}
\format{A character vector of probe names with length 120009}
\source{
Edgar R.D., Jones M.J., Robinson W.P., Kobor M.S. (2017) An 
empirically driven data reduction method on the human 450K methylation array 
to remove tissue specific non-variable CpGs. Clin. Epigenet ., 9, 11.
https://github.com/redgar598/Tissue_Nonvariable_450K_CpGs
}
\usage{
invar_Bu
}
\description{
A character vector of 120009 probe names identified as nonvariable between
multiple BUCCAL datasets.
}
\keyword{datasets}
