% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invar_Pl}
\alias{invar_Pl}
\title{Probes identified as invariable in indepedent placenta datasets}
\format{A character vector of probe names with length 101367}
\source{
Edgar R.D., Jones M.J., Robinson W.P., Kobor M.S. (2017) An 
empirically driven data reduction method on the human 450K methylation array 
to remove tissue specific non-variable CpGs. Clin. Epigenet ., 9, 11.
https://github.com/redgar598/Tissue_Nonvariable_450K_CpGs
}
\usage{
invar_Pl
}
\description{
A character vector of 101367 probe names identified as nonvariable between
multiple PLACENTA datasets.
}
\keyword{datasets}
