% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmatrix.R
\name{lmmatrix}
\alias{lmmatrix}
\title{Creates a matrix of p values or rsquared between a matrix of independent
variables and a dataframe of dependent variables}
\usage{
lmmatrix(dep, ind, metric = "Rsquared")
}
\arguments{
\item{dep}{A d x n matrix of dependent variables, where the variables are
arranged in rows, and columns for samples or observations.}

\item{ind}{A i x n dataframe of independent variables, in same format as dep.}

\item{metric}{Can be 'Rsquared' or 'Pvalue'. Determines contents of output
matrix}
}
\value{
A i x d matrix of pvalues or rsquared values.
}
\description{
\code{heatPCA} tests association between independent variables and dependent
variables through simple linear regression (1 predictor variable)
}
\details{
Each independent variable is tested for their association with each
dependent variable in simple linear regression, and a pvalue or rsquared is
extracted and returned as a matrix.
}
\examples{
## to calculate PC association with covariates

library(minfiData)
data(RGsetEx)
betas <- na.omit(getBeta(RGsetEx))
pDat <- as.data.frame(pData(RGsetEx))
pc_obj <- prcomp(t(betas))
pc_matrix <- t(pc_obj$x)
cov <- t(pDat[,c('Sample_Group', 'age', 'sex')])

lmmatrix(dep = pc_matrix, ind = cov)
lmmatrix(dep = pc_matrix, ind = cov, metric = 'Pvalue')
}
