% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSampleSheet.R
\name{makeSampleSheet}
\alias{makeSampleSheet}
\title{Creates a sample sheet out of the Robinson Lab's master sample sheet}
\usage{
makeSampleSheet(samples, ss_path = NULL, data_path = NULL)
}
\arguments{
\item{samples}{a vector of sample IDs}

\item{ss_path}{file path to master sample sheet. Defaults to 'Z:/ROBLAB6 Infinium450k John/450k 
Sample Sheets/MASTER SS.xlsx'}

\item{data_path}{file path to a parent directory containing idats. Idats can be in 
subdirectories within the parent directory. Defaults to 'Z:/ROBLAB6 Infinium450k John/450k Raw 
Data'}
}
\description{
\code{makeSampleSheet} find idats and returnsa data frame for a given set of samples.
}
\details{
DOES NOT WORK CURRENTLY, NEED TO UPDATE WITH NEW MASTER SS

Takes a vector of sample IDs and finds relevant experiment information. For example, 
plate, batch, Chip ID (Sentrix_ID), Row and column (Sentrix Position), and full file path are 
returned.

First it finds the sentrix Id and position from the master sample sheet (currently 450k only), 
then it finds the associated idats within the 450k Raw data folder, and returns file path 
information. If the robinson lab shared drive is mapped to a drive other than "Z:/", then you can
specify change the file paths with `ss_path` and `data_path`.

If a sample has been ran more than once, a dataframe will be printed along with batch / tissue 
information.

If there are copies of the same idats located in different directories, both will be added to the
data frame. Be aware that multiple copies of the same idats won't be able to be loaded with 
minfi::read.metharray.sheet.
}
\examples{
samples <- c('PM4', 'PM30', 'PM47', 'PM123', 'PM130', 'PM252')
ss <- makeSampleSheet(samples)
rgset <- read.metharray.exp(targets = ss, verbose = T)

}
