% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mprice_CH}
\alias{mprice_CH}
\title{List of XY- and autosomal cross-hybridizing probes from m.price et al. (2013)}
\format{A character vector of CpG names with length 41937}
\source{
Chen Y, Lemire M, Choufani S, et al. Discovery of cross-reactive 
probes and polymorphic CpGs in the Illumina Infinium HumanMethylation450 
microarray. Epigenetics. 2013;8(2):203-209. doi:10.4161/epi.23470.

https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL16304
}
\usage{
mprice_CH
}
\description{
A character vector of 41937 CpG names that were identified to be able to cross
hybridize to other locations in the genome (both XY and autosomal).
}
\keyword{datasets}
