% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overlap_450k_EPIC}
\alias{overlap_450k_EPIC}
\title{Probes that are on both the 450K and EPIC arrays}
\format{A character vector of probe names with length 453152}
\source{
minfiData 0.24.0, minfiDataEPIC 1.4.0
}
\usage{
overlap_450k_EPIC
}
\description{
A character vector of 453152 probes names that are on both the 450k and EPIC 
Illummina DNA methylation microarrays. All of the snp probes on the EPIC (59)
are included - only 6 are missing from the 450k array. These names were taken
from minfiData and minfiDataEPIC, because the illumina annotations do not have
snp probe information.
}
\keyword{datasets}
