% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairtest.R
\name{pairtest}
\alias{pairtest}
\title{Tests if every column is independent of each other}
\usage{
pairtest(df)
}
\arguments{
\item{df}{a dataframe of variables, where the variables are arranged in 
columns, and samples or observations in rows.}
}
\description{
\code{pairtest} tests association between columns of a dataframe.
}
\details{
Each variable is tested for their association with each other 
variable in simple linear regression (if at least one numerical variable), 
or chi squared test (if two categorical).
}
\examples{
## to calculate PC association with covariates

library(plyr)

variables <- data.frame(
  row = c(1, 1, 2, 2, 3, 3),
  column = c(1, 1, 1, 2, 2, 2),
  Sex = c('m', 'm', 'm', 'f', 'f', 'f'),
  age = c(18, 19, 18, 27, 30, 16),
  ethnicity = c('AF', 'AF', 'AF', 'EU', 'EU', 'AS'))
  
pairtest(variables)
}
