% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plomics-package.R
\docType{package}
\name{plomics-package}
\alias{plomics}
\alias{plomics-package}
\title{plomics: Miscellaneous functions, mainly for DNAme analysis}
\description{
This package contains functions that I have developed mainly for
    analysis of DNA methylation array data (450k, EPIC).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/wvictor14/yahew}
  \item Report bugs at \url{http://github.com/wvictor14/yahew}
}

}
\keyword{internal}
