% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess methylation array data}
\usage{
preprocess(mset, fp = NULL, ch = NULL, invariable_probes = NULL,
  pkeep = NULL, overlapping = NULL, wrongsex, seed = 1)
}
\arguments{
\item{mset}{A minfi::methylset object.}

\item{fp}{An p by n matrix/dataframe, where n = the number of samples and p =
number of features. This should be the same dimensions as mset, unless snp
probes are included in mset.}

\item{ch}{A character vector of cross hybridizing probes to be filtered out.}

\item{invariable_probes}{A character vector of reference invariable probes to
filter against.}

\item{overlapping}{A character vector of overlapping probes on 450k/EPIC
array to keep.}

\item{wrongsex}{A character vector of samplenames to filter out.}

\item{seed}{Sets a seed for normalization.}
}
\value{
List of length two:
\itemize{
  \item data - A filtered and normalized dataframe
  \item pDat - A dataframe containing updated phenotype data information
}
}
\description{
\code{preprocess} returns a list of length two, containing (1) filtered,
normalized dataframe and (2) pData with QC information
}
\details{
This function takes in a methylset object, and does the following:
(1) Filters poor quality probes based on fp, (2) normalizes beta values with
wateRmelon::BMIQ, (3) filters overlapping 450k/EPIC probes and invariable
probes. FP needs to be precomputed. Any of these probes can be omitted.

The reason of filtering overlapping / invariable probes is that normalization
can benefit from these probes as extra observations, whereas poor quality 
probes are removed prior to normalization.

If pkeep is supplied, then every other probe setting is overridden, and only
probes in pkeep will be kept, this filtering step is done after 
normalization.
}
