% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmatrix.R
\name{lmmatrix}
\alias{lmmatrix}
\title{Tests associations between multiple independent and dependent variables}
\usage{
lmmatrix(dep, ind, metric = "Rsquared")
}
\arguments{
\item{dep}{n x d matrix or dataframe of dependent variables, where the
variables are arranged in columns, and samples or observations in rows.}

\item{ind}{n x i matrix or dataframe of independent variables, where the
variables are arranged in columns, and samples or observations in rows.}

\item{metric}{Determines what is returned from the linear models. 'Rsquared' 
returns the proportion of variance explained, and 'Pvalue' returns the 
p-value.}
}
\value{
A i x d matrix of pvalues or rsquared values.
}
\description{
\code{lmmatrix} tests association between independent variables and dependent
variables through simple linear regression (with 1 predictor variable)
}
\details{
Each independent variable is tested for their association with each
dependent variable in simple linear regression (ind ~ dep), and a pvalue or
rsquared is extracted and returned as a matrix.
}
\examples{
## to calculate PC association with covariates

library(minfiData)
data(RGsetEx)
betas <- na.omit(getBeta(RGsetEx))
pDat <- as.data.frame(pData(RGsetEx))
pc_obj <- prcomp(t(betas))
pc_matrix <- pc_obj$x
cov <- pDat[,c('Sample_Group', 'age', 'sex')]

lmmatrix(dep = pc_matrix, ind = cov)
lmmatrix(dep = pc_matrix, ind = cov, metric = 'Pvalue')
}
