% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmatrix.R
\name{lmmatrix}
\alias{lmmatrix}
\title{Creates a matrix of p values or rsquared between a matrix of independent
variables and a dataframe of dependent variables}
\usage{
lmmatrix(dep, ind, metric = "Rsquared")
}
\arguments{
\item{dep}{A p x n matrix of dependent variables, where p are the variables
and n are observations or samples.}

\item{ind}{A n x m dataframe, where m columns are the number of covariates
to be tested for association with the dependent variables}

\item{metric}{Can be 'Rsquared' or 'Pvalue'. Determines contents of output
matrix}
}
\value{
A p x m matrix of pvalues or rsquared values.
}
\description{
\code{heatPCA} tests association between independent variables and dependent
variables through simple linear regression (1 predictor variable)
}
\details{
Each independent variable is tested for their association with each
dependent variable in simple linear regression, and a pvalue or rsquared is
extracted and returned as a matrix.
}
\examples{
## to calculate PC association with covariates

library(minfiData)
data(RGsetEx)
betas <- na.omit(getBeta(RGsetEx))
pDat <- pData(RGsetEx)
pc_obj <- prcomp(t(betas))
pc_matrix <- pc_obj$x

lmmatrix(dep = pc_matrix, ind = pDat[,c('Sample_Group', 'age', 'sex')])
lmmatrix(dep = pc_matrix, ind = pDat[,c('Sample_Group', 'age', 'sex')],
         metric = 'Pvalue')
}
