% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess methylation array data}
\usage{
preprocess(mset, fp = NULL, ch = NULL, invariable_probes = NULL,
  threshold_p = 0.01, threshold_s = 0.05, threshold_cor = NULL,
  overlapping = F, wrongsex = NULL, seed = 1)
}
\arguments{
\item{mset}{A minfi::methylset object.}

\item{fp}{An p by n matrix/dataframe, where n = the number of samples and p =
number of features. This should be the same dimensions as mset, unless snp
probes are included in mset.}

\item{ch}{A character vector specificying the set of cross hybridizing probes
to filter out. Options: 'mprice', 'chen', 'both'.}

\item{invariable_probes}{A character vector of specifying the set of
invariable probes to filter out. Options: 'Blood', 'Buccal', 'Placenta'.}

\item{threshold_p}{A numeric value between 0 and 1 representing the fraction
of observations required for a probe to be removed.}

\item{threshold_s}{A numeric value between 0 and 1 representing the fraction
of observations required for a sample to be removed.}

\item{threshold_cor}{A numeric value between 0 and 1 specifying the minimum
mean interarray correlation to keep a sample. If not supplied, then
correlations are not calculated.}

\item{overlapping}{A logical vector indicating if 450k/EPIC overlapping
probes should be kept.}

\item{wrongsex}{A character vector of samplenames to filter out.}

\item{seed}{Sets a seed for normalization.}
}
\value{
List of length two:
\itemize{
  \item data - A filtered and normalized tibble of betas
  \item pDat - A tible containing updated phenotype data information
  \item num_fp - number of probes filtered based on quality of signal
  \item num_ch - number of cross hybridizing probes filtered
  \item num_iv - number of invariable probes filtered
}
}
\description{
\code{preprocess} returns a list of length two, containing (1) filtered,
normalized tibble and (2) a pData tibble with QC information
}
\details{
This function takes in a methylset object, and does the following:
(1) Filters poor quality probes based on fp, (2) normalizes beta values with
wateRmelon::BMIQ, (3) filters overlapping 450k/EPIC probes and invariable
probes. FP needs to be precomputed. Any of these probes can be omitted.

The default for probe filtering is 0.01, and the default for sample filtering
is 0.05.

The reason of filtering overlapping / invariable probes is that normalization
can benefit from these probes as extra observations, whereas poor quality
probes are removed prior to normalization.
}
